# cache: Improve jump hash server selector performance