**What this PR does**:

Improve the performance of the jump hash Memcached ServerSelector implementation by _not_ delegating to the default implementation in `grafana/gomemcache`. This avoids the need to iterate through the entire list of servers to pick a single address.

I've tested the new implementation along side the original and the results (which server was picked for which key) were identical. Performance is about 3-4X better.

**Which issue(s) this PR fixes**:

See https://github.com/grafana/gomemcache/pull/22

**Checklist**
- [X] Tests updated
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`